% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyze-roc.R
\name{auc}
\alias{auc}
\title{Function \code{auc}}
\usage{
auc(roc, cutoff = 0.1)
}
\arguments{
\item{roc}{data frame with columns named \code{fpr} and \code{tpr}}

\item{cutoff}{fpr cutoff below which area will be calculated}
}
\value{
area under an ROC curve
}
\description{
calculates areas under ROC curves
}

