% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze-auc_df.R
\name{auc_df}
\alias{auc_df}
\title{Function \code{auc_df}}
\usage{
auc_df(path)
}
\arguments{
\item{path}{to directory of extracted results files}
}
\description{
make a auc data frame out of rds files extracted from simulation lists
}

