% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze-auc_long.R
\name{auc_long}
\alias{auc_long}
\title{Function \code{auc_long}}
\usage{
auc_long(from, to)
}
\arguments{
\item{from}{directory of extracted results files}

\item{to}{directory to save plots}
}
\description{
plot areas under roc curves using rds files extracted from simulation lists
}

