% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/benchmarks.R
\name{benchmarks}
\alias{benchmarks}
\title{Function \code{benchmarks}}
\usage{
benchmarks(path, methods = c("edgeR", "Niemi"), ncores = 1)
}
\arguments{
\item{path}{to directory to save simulations and results}

\item{methods}{methods to run}

\item{ncores}{number of cores to use}
}
\value{
a list if pertinent scenario information
}
\description{
run benchmark methods on simulation objects
}

