% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze-calibration.R
\name{calibration}
\alias{calibration}
\title{Function \code{calibration}}
\usage{
calibration(probs, truth)
}
\arguments{
\item{probs}{probabilities}

\item{truth}{logical or 0/1 vector of classifications}
}
\value{
data frame with an calibration curve
}
\description{
assess calibration of posterior probabilities
}

