% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze-calibration_long.R
\name{calibration_long}
\alias{calibration_long}
\title{Function \code{calibration_long}}
\usage{
calibration_long(from, to)
}
\arguments{
\item{from}{directory with calibration information}

\item{to}{directory to save plots}
}
\description{
plot calibration curves using rds files extracted from simulation lists
}

