% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyze-calibration.R
\name{calibration_over}
\alias{calibration_over}
\title{Function \code{calibration_over}}
\usage{
calibration_over(l, a)
}
\arguments{
\item{l}{simulation list}

\item{a}{analysis list}
}
\description{
use with \code{over_ans} to calculate calibration curves on analysis elements of simulation lists
}

