% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyze-calibration.R
\name{calibrations}
\alias{calibrations}
\title{Function \code{calibrations}}
\usage{
calibrations(from, to)
}
\arguments{
\item{from}{directory with simulation lists}

\item{to}{output directory}
}
\description{
Compute calibration curves for every \code{analyses} element of every simulation list in a directory
}

