% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/case.R
\name{case}
\alias{case}
\title{Function \code{case}}
\usage{
case()
}
\arguments{
\item{path}{to directory to save simulations and results}
}
\value{
path to simulated objects
}
\description{
Whole orkflow for the part of the case study paper that looks at gene detection power
}

