% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/case_coverage.R
\name{case_coverage}
\alias{case_coverage}
\title{Function \code{case_coverage}}
\usage{
case_coverage(path = newdir())
}
\arguments{
\item{path}{to directory to save simulations and results}
}
\value{
path to simulated objects
}
\description{
Workflow for the part of the case study paper that looks at the coverage of credible intervals
}

