% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/case_coverage_analyze.R
\name{case_coverage_analyze}
\alias{case_coverage_analyze}
\title{Function \code{case_coverage_analyze}}
\usage{
case_coverage_analyze(path = newdir())
}
\arguments{
\item{path}{to directory to save simulations and results}
}
\value{
path to simulated objects
}
\description{
Plot and summarize results from simulation study
}

