% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/case_coverage_fit.R
\name{case_coverage_fit}
\alias{case_coverage_fit}
\title{Function \code{case_coverage_fit}}
\usage{
case_coverage_fit(path = newdir())
}
\arguments{
\item{path}{to directory to save simulations and results}
}
\value{
path to simulated objects
}
\description{
Fit for part of the case study paper that looks at the coverage of credible intervals
}

