% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/case_coverage_init.R
\name{case_coverage_init}
\alias{case_coverage_init}
\title{Function \code{case_coverage_init}}
\usage{
case_coverage_init(path = newdir())
}
\arguments{
\item{path}{to directory to save simulations and results}
}
\value{
path to simulated objects
}
\description{
Initialize the part of the case study paper that looks at the coverage of credible intervals
}

