% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/case.R, R/case_power.R
\name{case_power}
\alias{case_power}
\title{Function \code{case}}
\usage{
case_power(path = newdir())

case_power(path = newdir())
}
\arguments{
\item{path}{to directory to save simulations and results}

\item{path}{to directory to save simulations and results}
}
\value{
path to simulated objects

path to simulated objects
}
\description{
Whole orkflow for the part of the case study paper that looks at gene detection power

Workflow for the part of the case study paper that looks at gene detection power
}

