% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/case_power_fit.R
\name{case_power_fit}
\alias{case_power_fit}
\title{Function \code{case_power_fit}}
\usage{
case_power_fit(path = newdir())
}
\arguments{
\item{path}{to directory to save simulations and results}
}
\value{
path to simulated objects
}
\description{
Fit for part of the case study paper that looks at gene detection power
}

