% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/case_study.R
\name{case_study}
\alias{case_study}
\title{Function \code{case_study}}
\usage{
case_study(path = newdir())
}
\arguments{
\item{path}{to directory to save simulations and results}
}
\value{
path to simulated objects
}
\description{
workflow of the case study paper
}

