% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze-ci.R
\name{ci}
\alias{ci}
\title{Function \code{ci}}
\usage{
ci(from, to, levels = c(0.5, 0.95))
}
\arguments{
\item{from}{directory with simulation lists}

\item{to}{where to put ci information}

\item{levels}{credible levels}
}
\value{
Info to assesses coverage of beta parameters by credible intervals.
}
\description{
Info to assesses coverage of beta parameters by credible intervals.
}

