% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyze-ci_beta.R
\name{ci_beta}
\alias{ci_beta}
\title{Function \code{ci_beta}}
\usage{
ci_beta(from, to)
}
\arguments{
\item{from}{directory with simulation _levels}

\item{to}{where to put ci information}
}
\value{
Info to assesses coverage of betaparameters by credible intervals.
}
\description{
Info to assesses coverage of betaparameters by credible intervals.
}

