% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyze-ci_hyper.R
\name{ci_hyper}
\alias{ci_hyper}
\title{Function \code{ci_level}}
\usage{
ci_hyper(from, to)
}
\arguments{
\item{from}{directory with simulation lists}

\item{to}{where to put ci information}
}
\value{
Info to assesses coverage of hyperparameters by credible intervals.
}
\description{
Info to assesses coverage of hyperparameters by credible intervals.
}

