% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-clean_df.R
\name{clean_df}
\alias{clean_df}
\title{Function \code{clean_df}}
\usage{
clean_df(d)
}
\arguments{
\item{d}{raw data frame}
}
\value{
clean data frame for plotting
}
\description{
clean data frame for plotting
}

