% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze-comparehprobs.R
\name{comparehprobs}
\alias{comparehprobs}
\title{Function \code{comparehprobs}}
\usage{
comparehprobs(from, to)
}
\arguments{
\item{from}{directory containing mcmc results on real data}

\item{to}{directory to save plots}
}
\description{
plot information about coverage of beta parameters in credible intervals
}

