% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/study-comparison.R
\name{comparison}
\alias{comparison}
\title{Function \code{comparison}}
\usage{
comparison(path = newdir())
}
\arguments{
\item{path}{to directory to save simulations and results}
}
\value{
path to simulated objects
}
\description{
Workflow for the comparison study paper
}

