% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study-comparison_analyze.R
\name{comparison_analyze}
\alias{comparison_analyze}
\title{Function \code{comparison_analyze}}
\usage{
comparison_analyze(from, to)
}
\arguments{
\item{from}{to directory to save simulations and results}

\item{to}{output directory}
}
\description{
Plot and summarize results from simulation studies
}

