% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/comparison_fit.R
\name{comparison_fit}
\alias{comparison_fit}
\title{Function \code{comparison_fit}}
\usage{
comparison_fit(path = newdir())
}
\arguments{
\item{path}{to directory to save simulations and results}
}
\value{
path to simulated objects
}
\description{
Fit for part of the case study paper that looks at gene detection comparison
}

