% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/study-comparison_init.R
\name{comparison_init}
\alias{comparison_init}
\title{Function \code{comparison_init}}
\usage{
comparison_init(path = newdir())
}
\arguments{
\item{path}{to directory to save simulations and results}
}
\value{
path to simulated objects
}
\description{
Initialize the part of the case study paper that looks at gene detection comparison
}

