% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/study-comparison_mcmc.R
\name{comparison_mcmc}
\alias{comparison_mcmc}
\title{Function \code{comparison_mcmc}}
\usage{
comparison_mcmc(path = newdir())
}
\arguments{
\item{path}{to directory to save simulations and results}
}
\value{
path to simulated objects
}
\description{
MCMC for the comparison study
}

