% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/study-computation.R
\name{computation}
\alias{computation}
\title{Function \code{computation}}
\usage{
computation(path = newdir())
}
\arguments{
\item{path}{to directory to save simulations and results}
}
\value{
path to simulated objects
}
\description{
workflow of the computation study paper
}

