% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/study-computation_analyze.R
\name{computation_analyze}
\alias{computation_analyze}
\title{Function \code{computation_analyze}}
\usage{
computation_analyze(from, to)
}
\arguments{
\item{from}{to directory to save simulations and results}

\item{to}{output directory}
}
\description{
Plot and summarize results from simulation study
}

