% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/computation_fit.R
\name{computation_fit}
\alias{computation_fit}
\title{Function \code{computation_fit}}
\usage{
computation_fit(path = newdir())
}
\arguments{
\item{path}{to directory to save simulations and results}
}
\value{
path to simulated objects
}
\description{
Run the main methods on the simulated data
}

