% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/study-computation_init.R
\name{computation_init}
\alias{computation_init}
\title{Function \code{computation_init}}
\usage{
computation_init(path = newdir())
}
\arguments{
\item{path}{to directory to save simulations and results}
}
\value{
path to simulated objects
}
\description{
create the data files for the computation study
}

