% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study-computation_mcmc.R
\name{computation_mcmc}
\alias{computation_mcmc}
\title{Function \code{computation_mcmc}}
\usage{
computation_mcmc(path = newdir())
}
\arguments{
\item{path}{to directory to save simulations and results}
}
\value{
path to simulated objects
}
\description{
MCMC of the computation simulation study
}

