% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/computation_paper.R
\name{computation_paper_runtimes}
\alias{computation_paper_runtimes}
\title{Function \code{computation_paper_runtimes}}
\usage{
computation_paper_runtimes(path)
}
\arguments{
\item{path}{to directory to save simulations and results}
}
\value{
path to simulated objects
}
\description{
plot runtimes for computation paper
}

