% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/computation_study.R
\name{computation_study}
\alias{computation_study}
\title{Function \code{computation_study}}
\usage{
computation_study(path = newdir(), genes = c(8192, 16384, 32768),
  libraries = c(16, 32, 64), reps = 5)
}
\arguments{
\item{path}{to directory to save simulations and results}

\item{genes}{number of genes. Can be a vector.}

\item{libraries}{number of libraries. Can be a vector.}

\item{reps}{number of reps for each #genes/#libraries combination. Should be a scalar.}
}
\value{
path to simulated objects
}
\description{
workflow of the computation study paper
}

