% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/computation_study.R
\name{computation_study_gelman}
\alias{computation_study_gelman}
\title{Function \code{computation_study_gelman}}
\usage{
computation_study_gelman(path)
}
\arguments{
\item{path}{to directory to save simulations and results}
}
\value{
path to simulated objects
}
\description{
list of gelman factors
}

