% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/computation_study.R
\name{computation_study_runtime_plots}
\alias{computation_study_runtime_plots}
\title{Function \code{computation_study_runtime_plots}}
\usage{
computation_study_runtime_plots(path)
}
\arguments{
\item{path}{to directory to save simulations and results}
}
\value{
path to simulated objects
}
\description{
plot runtimes for computation paper
}

