% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/computation_study.R
\name{computation_study_runtime_table}
\alias{computation_study_runtime_table}
\title{Function \code{computation_study_runtime_table}}
\usage{
computation_study_runtime_table(path)
}
\arguments{
\item{path}{to directory to save simulations and results}
}
\value{
path to simulated objects
}
\description{
table of runtimes for computation paper
}

