% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/study-coverage.R
\name{coverage}
\alias{coverage}
\title{Function \code{coverage}}
\usage{
coverage(path = newdir())
}
\arguments{
\item{path}{to directory to save simulations and results}
}
\value{
path to simulated objects
}
\description{
Workflow for the part of the case study paper that looks at the coverage of credible intervals
}

