% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/study-coverage_analyze.R
\name{coverage_analyze}
\alias{coverage_analyze}
\title{Function \code{coverage_analyze}}
\usage{
coverage_analyze(path = newdir())
}
\arguments{
\item{path}{to directory to save simulations and results}
}
\value{
path to simulated objects
}
\description{
Plot and summarize results from simulation study
}

