% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study-coverage_init.R
\name{coverage_init}
\alias{coverage_init}
\title{Function \code{coverage_init}}
\usage{
coverage_init(path = newdir())
}
\arguments{
\item{path}{to directory to save simulations and results}
}
\value{
path to simulated objects
}
\description{
Initialize the part of the case study paper that looks at the coverage of credible intervals
}

