% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study-coverage_mcmc.R
\name{coverage_mcmc}
\alias{coverage_mcmc}
\title{Function \code{coverage_mcmc}}
\usage{
coverage_mcmc(path = newdir(), zeronormfactors = T)
}
\arguments{
\item{path}{to directory to save simulations and results}

\item{zeronormfactors}{TRUE/FALSE value: option to set normalization constants h to 0}
}
\value{
path to simulated objects
}
\description{
MCMC of the coverage study (credible intervals)
}

