% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-logfiles.R
\name{default_logfiles}
\alias{default_logfiles}
\title{Function \code{default_logfiles}}
\usage{
default_logfiles()
}
\value{
vector of paths path to log files
}
\description{
randomly-generated log file locations to \code{sink()} and delete verbose messages
}

