% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyze-ess.R
\name{ess}
\alias{ess}
\title{Function \code{ess}}
\usage{
ess(from, to)
}
\arguments{
\item{from}{to directory to save simulations and results}

\item{to}{output directory}
}
\description{
list effective sample sizes
}

