% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyze-explore_real.R
\name{explore_real}
\alias{explore_real}
\title{Function \code{explore_real}}
\usage{
explore_real(from, to)
}
\arguments{
\item{from}{directory with simulation lists}

\item{to}{output directory}
}
\description{
explore_reals in computation study
}

