% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyze-fdr.R
\name{fdr}
\alias{fdr}
\title{Function \code{fdr}}
\usage{
fdr(probs, truth)
}
\arguments{
\item{probs}{probabilities}

\item{truth}{logical or 0/1 vector of classifications}
}
\value{
data frame with an fdr curve
}
\description{
function for calculating fdr curves
}

