% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyze-fdr.R
\name{fdr_over}
\alias{fdr_over}
\title{Function \code{fdr_over}}
\usage{
fdr_over(l, a)
}
\arguments{
\item{l}{simulation list}

\item{a}{analysis list}
}
\description{
use with \code{over_ans} to calculate fdr curves on analysis elements of simulation lists
}

