% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyze-fdr.R
\name{fdrs}
\alias{fdrs}
\title{Function \code{fdrs}}
\usage{
fdrs(from, to)
}
\arguments{
\item{from}{directory with simulation lists}

\item{to}{output directory}
}
\description{
Compute fdr curves for every \code{analyses} element of every simulation list in a directory
}

