% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit_edgeR.R
\name{fit_edgeR}
\alias{fit_edgeR}
\title{Function \code{fit_edgeR}}
\usage{
fit_edgeR(counts, design)
}
\arguments{
\item{counts}{A data frame or matrix of RNA-seq read counts.}

\item{design}{design matrix}
}
\value{
a list of components of a fit object from \code{edgeR}
}
\description{
fits edgeR
}

