% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyze-roc.R
\name{fpr_point}
\alias{fpr_point}
\title{Function \code{fpr_point}}
\usage{
fpr_point(cutoff, probs, n)
}
\arguments{
\item{cutoff}{cutoff for fpr calculation}

\item{probs}{probabilities}

\item{n}{number of probabilities, not necessarily in \code{probs}}
}
\value{
false positive rate
}
\description{
internal function for computing ROC curves
}

