% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze-gelman.R
\name{gelman}
\alias{gelman}
\title{Function \code{gelman}}
\usage{
gelman(from, to)
}
\arguments{
\item{from}{to directory to save simulations and results}

\item{to}{output directory}
}
\description{
list of gelman factors
}

