% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_Niemi.R
\name{get_hyperparameters}
\alias{get_hyperparameters}
\title{Function \code{get_hyperparameters}}
\usage{
get_hyperparameters(counts, design)
}
\arguments{
\item{counts}{RNA-seq count dataset}

\item{design}{design matrix}
}
\value{
internal object
}
\description{
internal function
}

