% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting-ggplot2_df.R
\name{ggplot2_df}
\alias{ggplot2_df}
\title{Function \code{ggplot2_df}}
\usage{
ggplot2_df(path)
}
\arguments{
\item{path}{to directory of extracted results files}
}
\description{
make a ggplot2 data frame out of rds files extracted from simulation lists
}

