% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landau_dissertation.R
\name{Landau_dissertation}
\alias{Landau_dissertation}
\title{Function \code{Landau_dissertation}}
\usage{
Landau_dissertation()
}
\description{
Reproduce all the computation, figures, tables, etc. of the Statistics PhD
dissertation of Will Landau (http://will-landau.com, will.landau@gmail.com).
}

