% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze-mse.R
\name{mse}
\alias{mse}
\title{Function \code{mse}}
\usage{
mse(from, to)
}
\arguments{
\item{from}{directory with simulation lists}

\item{to}{output directory}
}
\value{
mean squared error
}
\description{
function for calculating mean squared error
}

