% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util-myrelevel.R
\name{myrelevel}
\alias{myrelevel}
\title{Function \code{myrelevel}}
\usage{
myrelevel(x)
}
\arguments{
\item{x}{factor to relevel}
}
\value{
releveled factor
}
\description{
relevel analysis vector for plotting labels
}

