% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-logfiles.R
\name{mysink}
\alias{mysink}
\title{Function \code{mysink}}
\usage{
mysink()
}
\value{
a list of logfiles and the file connection of the message file
}
\description{
sink to randomly-generated log file locations
}

