% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{newdir}
\alias{newdir}
\title{Function \code{newdir}}
\usage{
newdir(path = dir_name())
}
\arguments{
\item{path}{path to directory}
}
\value{
directory name
}
\description{
make a new directory for storing simulations and results
}

