% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/path.R
\name{newpath}
\alias{newpath}
\title{Function \code{newpath}}
\usage{
newpath(path = pathname())
}
\arguments{
\item{path}{path to directory}
}
\value{
directory name
}
\description{
make a new directory for storing simulations and results
}

