% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study-OpenMP_runs.R
\name{OpenMP_runs}
\alias{OpenMP_runs}
\title{Function \code{OpenMP_runs}}
\usage{
OpenMP_runs(path = newdir())
}
\arguments{
\item{path}{to directory to save simulations and results}
}
\value{
path to results
}
\description{
Execute OpenMP MCMC on Paschold data.
}

