% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-over.R
\name{over_out}
\alias{over_out}
\title{Function \code{over_out}}
\usage{
over_out(from, to, fun)
}
\arguments{
\item{from}{directory with simulation lists}

\item{to}{output directory}

\item{fun}{function to apply over every analysis-specific output file in a given directory}
}
\description{
apply a function over every analysis-specific output file in a given directory
}

