% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util-over.R
\name{over_sims}
\alias{over_sims}
\title{Function \code{over_sims}}
\usage{
over_sims(from, to, fun)
}
\arguments{
\item{from}{directory with simulation lists}

\item{to}{output directory}

\item{fun}{function to apply to each simulation list}
}
\description{
apply a function over simulation lists and save to another directory
}

