% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper_case_figures.R
\name{paper_case_figures}
\alias{paper_case_figures}
\title{Function \code{paper_case_figures}}
\usage{
paper_case_figures()
}
\description{
Reproduce the figures and tables of the case study paper
}

