% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper_priors_figures.R
\name{paper_priors_figures}
\alias{paper_priors_figures}
\title{Function \code{paper_priors_figures}}
\usage{
paper_priors_figures()
}
\description{
Reproduce the figures and tables of the case study paper
}

