% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyze-plot_auc.R
\name{plot_auc}
\alias{plot_auc}
\title{Function \code{plot_auc}}
\usage{
plot_auc(from, to)
}
\arguments{
\item{from}{directory of extracted results files}

\item{to}{directory to save plots}
}
\description{
plot areas under roc curves using rds files extracted from simulation lists
}

