% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyze-plot_calibration.R
\name{plot_calibration}
\alias{plot_calibration}
\title{Function \code{plot_calibration}}
\usage{
plot_calibration(from, to)
}
\arguments{
\item{from}{directory of extracted results files}

\item{to}{directory to save plots}
}
\description{
plot posterior probability calibration curves using rds files extracted from simulation lists
}

