% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze-plot_calibration_df.R
\name{plot_calibration_df}
\alias{plot_calibration_df}
\title{Function \code{plot_calibration_df}}
\usage{
plot_calibration_df(df, to, analysis = analyses(), reps = 1:10)
}
\arguments{
\item{df}{output of ggplot2_df}

\item{to}{directory to save plots}

\item{analysis}{analysis methods to plot}

\item{reps}{reps to plot}
}
\description{
plot posterior probability calibration curves using rds files extracted from simulation lists
}

