% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze-plot_ci.R
\name{plot_ci}
\alias{plot_ci}
\title{Function \code{plot_ci}}
\usage{
plot_ci(from, to)
}
\arguments{
\item{from}{directory containing ci information}

\item{to}{directory to save plots}
}
\description{
plot information about coverage of beta parameters in credible intervals
}

