% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyze-plot_ci_beta.R
\name{plot_ci_beta}
\alias{plot_ci_beta}
\title{Function \code{plot_ci_beta}}
\usage{
plot_ci_beta(from, to)
}
\arguments{
\item{from}{directory containing ci information}

\item{to}{directory to save plots}
}
\description{
plot roc curves using rds files extracted from simulation lists
}

