% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyze-plot_roc.R
\name{plot_roc}
\alias{plot_roc}
\title{Function \code{plot_roc}}
\usage{
plot_roc(from, to, cutoff = 0.1, analysis = analyses())
}
\arguments{
\item{from}{directory with roc information}

\item{to}{directory to save plots}

\item{cutoff}{for fpr}

\item{analysis}{analysis methods to plot}
}
\description{
plot roc curves using rds files extracted from simulation lists
}

