% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyze-plot_roc_df.R
\name{plot_roc_df}
\alias{plot_roc_df}
\title{Function \code{plot_roc_df}}
\usage{
plot_roc_df(df, to, cutoff = 0.1, analysis = analyses(), reps = 1:10)
}
\arguments{
\item{df}{output of ggplot2_df}

\item{to}{directory to save plots}

\item{cutoff}{for fpr}

\item{analysis}{analysis methods to plot}

\item{reps}{reps to plot}
}
\description{
plot roc curves using rds files extracted from simulation lists
}

