% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyze-plot_runtime.R
\name{plot_runtime}
\alias{plot_runtime}
\title{Function \code{plot_runtime}}
\usage{
plot_runtime(from, to)
}
\arguments{
\item{from}{directory with runtime table}

\item{to}{output directory}
}
\description{
Plots runtimes in computation study
}

