% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting-powerplots.R
\name{powerplots}
\alias{powerplots}
\title{Function \code{powerplots}}
\usage{
powerplots(from, to)
}
\arguments{
\item{from}{directory of extracted results files}

\item{to}{directory to save plots}
}
\description{
plot results in power study using rds files extracted from simulation lists
}

