% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/priors.R
\name{priors}
\alias{priors}
\title{Function \code{priors}}
\usage{
priors(path = newdir())
}
\arguments{
\item{path}{to directory to save simulations and results}
}
\value{
path to simulated objects
}
\description{
Workflow for the priors study paper
}

