% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/priors_analyze.R
\name{priors_analyze}
\alias{priors_analyze}
\title{Function \code{power_analyze}}
\usage{
priors_analyze(path = newdir())
}
\arguments{
\item{path}{to directory to save simulations and results}
}
\value{
path to simulated objects
}
\description{
Plot and summarize results from priors study
}

