% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/priors_fit.R
\name{priors_fit}
\alias{priors_fit}
\title{Function \code{priors_fit}}
\usage{
priors_fit(path = newdir())
}
\arguments{
\item{path}{to directory to save simulations and results}
}
\value{
path to simulated objects
}
\description{
Fit for part of the case study paper that looks at gene detection priors
}

