% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/priors_init.R
\name{priors_init}
\alias{priors_init}
\title{Function \code{priors_init}}
\usage{
priors_init(path = newdir())
}
\arguments{
\item{path}{to directory to save simulations and results}
}
\value{
path to simulated objects
}
\description{
Initialize the part of the case study paper that looks at gene detection priors
}

