% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study-priors_mcmc.R
\name{priors_mcmc}
\alias{priors_mcmc}
\title{Function \code{priors_mcmc}}
\usage{
priors_mcmc(path = newdir(), zeronormfactors = T)
}
\arguments{
\item{path}{to directory to save simulations and results}

\item{zeronormfactors}{TRUE/FALSE value: option to set normalization constants h to 0}
}
\value{
path to simulated objects
}
\description{
MCMC of the priors study (credible intervals)
}

