% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-relevel_analyses.R
\name{priors_relevel_analyses}
\alias{priors_relevel_analyses}
\title{Function \code{priors_relevel_analyses}}
\usage{
priors_relevel_analyses(x)
}
\arguments{
\item{x}{factor to relevel}
}
\value{
releveled factor
}
\description{
relevel analysis vector for plotting labels
}

