% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-progress.R
\name{progress}
\alias{progress}
\title{Function \code{progress}}
\usage{
progress(path)
}
\arguments{
\item{path}{path to simulated objects}
}
\value{
path to simulated objects
}
\description{
check on progress of simulation study
}

