% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/study-real.R
\name{real}
\alias{real}
\title{Function \code{real}}
\usage{
real(path = newdir())
}
\arguments{
\item{path}{to directory to save simulations and results}
}
\value{
path to simulated objects
}
\description{
workflow of the real data analysis
}

