% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/real_analyze.R
\name{real_analyze}
\alias{real_analyze}
\title{Function \code{real_analyze}}
\usage{
real_analyze(path = newdir())
}
\arguments{
\item{path}{to directory to save simulations and results}
}
\value{
path to simulated objects
}
\description{
Plot and summarize results from the real data analysis
}

