% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/real_fit.R
\name{real_fit}
\alias{real_fit}
\title{Function \code{real_fit}}
\usage{
real_fit(path = newdir())
}
\arguments{
\item{path}{to directory to save simulations and results}
}
\value{
path to simulated objects
}
\description{
Run the main methods on the real data analysis
}

