% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/study-real_init.R
\name{real_init}
\alias{real_init}
\title{Function \code{real_init}}
\usage{
real_init(path = newdir())
}
\arguments{
\item{path}{to directory to save simulations and results}
}
\value{
path to simulated objects
}
\description{
create the data files for the real data analysis
}

