% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/study-real_mcmc.R
\name{real_mcmc}
\alias{real_mcmc}
\title{Function \code{real_mcmc}}
\usage{
real_mcmc(path = newdir())
}
\arguments{
\item{path}{to directory to save simulations and results}
}
\value{
path to simulated objects
}
\description{
MCMC of the real data analysis
}

