% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util-relevel.R
\name{relevel}
\alias{relevel}
\title{Function \code{relevel}}
\usage{
relevel(x)
}
\arguments{
\item{x}{factor to relevel}
}
\value{
releveled factor
}
\description{
relevel analysis vector for plotting labels
}

