% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-relevel_analyses.R
\name{relevel_analyses}
\alias{relevel_analyses}
\title{Function \code{relevel_analyses}}
\usage{
relevel_analyses(x)
}
\arguments{
\item{x}{factor to relevel}
}
\value{
releveled factor
}
\description{
relevel analysis vector for plotting labels
}

