% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-relevel_heterosis.R
\name{relevel_heterosis}
\alias{relevel_heterosis}
\title{Function \code{relevel_heterosis}}
\usage{
relevel_heterosis(x)
}
\arguments{
\item{x}{factor to relevel}
}
\value{
releveled factor
}
\description{
relevel heterosis status vector for plotting labels
}

