% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-relevel_simulations.R
\name{relevel_simulations}
\alias{relevel_simulations}
\title{Function \code{relevel_simulations}}
\usage{
relevel_simulations(x)
}
\arguments{
\item{x}{factor to relevel}
}
\value{
releveled factor
}
\description{
relevel simulations vector for plotting labels
}

