% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/repair.R
\name{repairs}
\alias{repairs}
\title{Function \code{repair}}
\usage{
repairs(from, to)
}
\arguments{
\item{from}{directory with simulation lists}

\item{to}{output directory}
}
\description{
make repairs to names, etc. to simulation lists
}

