% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyze-roc.R
\name{roc}
\alias{roc}
\title{Function \code{roc}}
\usage{
roc(probs, truth, cutoffs = c(0.05, 0.1, 0.15, 0.2, 0.5, 1))
}
\arguments{
\item{probs}{probabilities}

\item{truth}{logical or 0/1 vector of classifications}

\item{cutoffs}{cutoffs for calculating the area below}
}
\value{
data frame with an roc curve
}
\description{
function for calculating ROC curves
}

