% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze-roc_long.R
\name{roc_long}
\alias{roc_long}
\title{Function \code{roc_long}}
\usage{
roc_long(from, to)
}
\arguments{
\item{from}{directory with roc information}

\item{to}{directory to save plots}
}
\description{
plot roc curves using rds files extracted from simulation lists
}

