% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze-roc.R
\name{roc_over}
\alias{roc_over}
\title{Function \code{roc_over}}
\usage{
roc_over(l, a)
}
\arguments{
\item{l}{simulation list}

\item{a}{analysis list}
}
\description{
use with \code{over_ans} to calculate ROC curves on analysis elements of simulation lists
}

