% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze-roc.R
\name{rocs}
\alias{rocs}
\title{Function \code{rocs}}
\usage{
rocs(from, to)
}
\arguments{
\item{from}{directory with simulation lists}

\item{to}{output directory}
}
\description{
Compute roc curves for every \code{analyses} element of every simulation list in a directory
}

