% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyze-runtime_plots.R
\name{runtime_plots}
\alias{runtime_plots}
\title{Function \code{runtime_plots}}
\usage{
runtime_plots(path)
}
\arguments{
\item{path}{to directory to save simulations and results}
}
\value{
path to simulated objects
}
\description{
plot runtimes for computation paper
}

