% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyze-runtime_tables.R
\name{runtime_tables}
\alias{runtime_tables}
\title{Function \code{runtime_tables}}
\usage{
runtime_tables(path)
}
\arguments{
\item{path}{to directory to save simulations and results}
}
\value{
path to simulated objects
}
\description{
table of runtimes for computation paper
}

