% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study-serial_mcmc.R
\name{serial_mcmc}
\alias{serial_mcmc}
\title{Function \code{serial_mcmc}}
\usage{
serial_mcmc(path = newdir())
}
\arguments{
\item{path}{to directory to save simulations and results}
}
\value{
path to results
}
\description{
Execute serial MCMC on Paschold data.
}

