% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study-serial_runs.R
\name{serial_runs}
\alias{serial_runs}
\title{Function \code{serial_runs}}
\usage{
serial_runs(path = newdir())
}
\arguments{
\item{path}{to directory to save simulations and results}
}
\value{
path to results
}
\description{
Execute serial MCMC on Paschold data.
}

