% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_duplicated.R
\name{simulation_duplicated}
\alias{simulation_duplicated}
\title{Function \code{simulation_duplicated}}
\usage{
simulation_duplicated(genes = 35000, libraries = 16)
}
\arguments{
\item{genes}{number of genes}

\item{libraries}{number of libraries}
}
\value{
a list if pertinent scenario information
}
\description{
simulation scenario made by simply duplicating the rows and columns of the Paschold data
}

