% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation_edgeR.R
\name{simulation_edgeR}
\alias{simulation_edgeR}
\title{Function \code{simulation_edgeR}}
\usage{
simulation_edgeR(genes = 30000, libraries = 16, fit = NULL)
}
\arguments{
\item{genes}{number of genes}

\item{libraries}{number of libraries}

\item{fit}{output object from fit_edgeR}
}
\value{
a list if pertinent scenario information
}
\description{
simulates a scenario using edgeR
}

