% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation_Niemi.R
\name{simulation_Niemi}
\alias{simulation_Niemi}
\title{Function \code{simulation_Niemi}}
\usage{
simulation_Niemi(genes = 30000, libraries = 16)
}
\arguments{
\item{genes}{number of genes}

\item{libraries}{number of libraries}
}
\value{
a list if pertinent scenario information
}
\description{
simulates a scenario using Niemi
}

