% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_paschold.R
\name{simulation_paschold}
\alias{simulation_paschold}
\title{Function \code{simulation_paschold}}
\usage{
simulation_paschold()
}
\value{
a list if pertinent scenario information
}
\description{
plugs the real Paschold dataset into the simulation workflow
}

