% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_priors.R
\name{simulation_priors}
\alias{simulation_priors}
\title{Function \code{simulation_priors}}
\usage{
simulation_priors(genes = 10000, libraries = 16, prior = "normal")
}
\arguments{
\item{genes}{number of genes}

\item{libraries}{number of libraries. Must be even.}

\item{prior}{on the beta_{2, g} parameters. one of "normal", "Laplace", or "t"}
}
\value{
a list if pertinent scenario information
}
\description{
simulates a small differential expression scenario
}

