% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation_simple.R
\name{simulation_simple}
\alias{simulation_simple}
\title{Function \code{simulation_simple}}
\usage{
simulation_simple(genes = 30000, libraries = 16)
}
\arguments{
\item{genes}{number of genes}

\item{libraries}{number of libraries}
}
\value{
a list if pertinent scenario information
}
\description{
simulates a scenario using a naive method
}

