% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyze-table_ci.R
\name{table_ci}
\alias{table_ci}
\title{Function \code{table_ci}}
\usage{
table_ci(from, to)
}
\arguments{
\item{from}{directory containing ci information}

\item{to}{directory to save plots}
}
\description{
table information about coverage of parameters in credible intervals
}

