% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyze-table_ci_beta.R
\name{table_ci_beta}
\alias{table_ci_beta}
\title{Function \code{table_ci_beta}}
\usage{
table_ci_beta(from, to)
}
\arguments{
\item{from}{directory containing ci information}

\item{to}{directory to save plots}
}
\description{
table information about coverage of beta parameters in credible intervals
}

