% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tableS3table1}
\alias{tableS3table1}
\title{tableS3table1}
\description{
a data frame object. Contains the "Table 1" section of Supplementary Table S3
of the Paschold et al. (2012) study (http://www.ncbi.nlm.nih.gov/pubmed/23086286).
}
\author{
Paschold et al. (http://www.ncbi.nlm.nih.gov/pubmed/23086286)
}

