% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util-logfiles.R
\name{unsink}
\alias{unsink}
\title{Function \code{unsink}}
\usage{
unsink(logs)
}
\arguments{
\item{logs}{list of files and the connection for output and messages}
}
\description{
unsink message and output. also remove previous logfiles.
}

