% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyze-ci_beta.R
\name{ci_beta}
\alias{ci_beta}
\title{Function \code{ci_beta}}
\usage{
ci_beta(from, to)
}
\arguments{
\item{from}{directory containing ci information}

\item{to}{directory to save plots}
}
\description{
plot roc curves using rds files extracted from simulation lists
}

