% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyze-ci_hyper.R
\name{ci_hyper}
\alias{ci_hyper}
\title{Function \code{ci_hyper}}
\usage{
ci_hyper(from, to)
}
\arguments{
\item{from}{directory containing ci information}

\item{to}{directory to save plots}
}
\description{
plot roc curves using rds files extracted from simulation lists
}

