% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyze-ci_hyper_list.R
\name{ci_hyper_list}
\alias{ci_hyper_list}
\title{Function \code{ci_hyper_list}}
\usage{
ci_hyper_list(from, to, level = 0.5)
}
\arguments{
\item{from}{directory with simulation lists}

\item{to}{where to put ci information}

\item{level}{credible level}
}
\value{
Info to assesses coverage of hyperparameters by credible intervals.
}
\description{
Info to assesses coverage of hyperparameters by credible intervals.
}

