% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/publish-comparehprobs.R
\name{comparehprobs}
\alias{comparehprobs}
\title{Function \code{comparehprobs}}
\usage{
comparehprobs(from, from_s3, to)
}
\arguments{
\item{from}{directory containing mcmc results on real data}

\item{from_s3}{directory containing classification information of genes according to Paschold (2012)}

\item{to}{directory to save plots}
}
\description{
plot information about coverage of beta parameters in credible intervals
}

