% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study-comparison_init.R
\name{comparison_init}
\alias{comparison_init}
\title{Function \code{comparison_init}}
\usage{
comparison_init(path = newdir(), ncores = detectCores())
}
\arguments{
\item{path}{to directory to save simulations and results}

\item{ncores}{number of cores for \code{fit_Niemi()}}
}
\value{
path to simulated objects
}
\description{
Initialize the part of the case study paper that looks at gene detection comparison
}

