% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/comparison_study.R
\name{comparison_study}
\alias{comparison_study}
\title{Function \code{comparison_study}}
\usage{
comparison_study(path = newdir())
}
\arguments{
\item{path}{to directory to save simulations and results}
}
\value{
path to simulated objects
}
\description{
workflow of the comparison study paper
}

