% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/computation_analyze.R
\name{computation_analyze}
\alias{computation_analyze}
\title{Function \code{computation_analyze}}
\usage{
computation_analyze(path = newdir())
}
\arguments{
\item{path}{to directory to save simulations and results}
}
\value{
path to simulated objects
}
\description{
Plot and summarize results from simulation study
}

