% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/computation_paper.R
\name{computation_paper}
\alias{computation_paper}
\title{Function \code{computation_paper}}
\usage{
computation_paper(path = newdir(), genes = c(1024, 4096, 16384),
  libraries = c(16, 32, 64), reps = 10)
}
\arguments{
\item{path}{to directory to save simulations and results}

\item{genes}{number of genes. Can be a vector.}

\item{libraries}{number of libraries. Can be a vector.}

\item{reps}{number of reps for each #genes/#libraries combination. Should be a scalar.}
}
\value{
path to simulated objects
}
\description{
workflow of the computation paper
}

