% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/study-coverage_mcmc.R
\name{coverage_mcmc}
\alias{coverage_mcmc}
\title{Function \code{coverage_mcmc}}
\usage{
coverage_mcmc(path = newdir())
}
\arguments{
\item{path}{to directory to save simulations and results}
}
\value{
path to simulated objects
}
\description{
MCMC of the coverage study (credible intervals)
}

