% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Function \code{fit}}
\usage{
fit(path, benchmarks = c("edgeR", "Niemi"), depths = c("fullybayes",
  "ebayes", "ebayes_from_truth"), priors = c("normal", alternate_priors()),
  ncores = 1, debug = F)
}
\arguments{
\item{path}{to directory to save simulations and results}

\item{benchmarks}{benchmark methods to run}

\item{depths}{either "fullybayes" or "ebayes" or a vector with both.}

\item{priors}{priors on the betas}

\item{ncores}{number of cores for CPU-parallel methods}

\item{debug}{debug mode, TRUE/FALSE}
}
\value{
path to simulated objects
}
\description{
run benchmark methods on simulation objects
}

