% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Function \code{fit}}
\usage{
fit(path, benchmarks = c("edgeR", "Niemi"), fbseq_methods = c("fullybayes",
  "ebayes", "ebayesFromTruth"), priors = c("normal", special_beta_priors()),
  ncores = detectCores(), debug = F, configs = Configs())
}
\arguments{
\item{path}{to directory to save simulations and results}

\item{benchmarks}{benchmark methods to run}

\item{fbseq_methods}{"fullybayes", "ebayes", "ebayes_from_truth" or a vector with a combination. "fullybayes" must come before "ebayes".}

\item{priors}{priors on the betas}

\item{ncores}{number of cores for CPU-parallel methods}

\item{debug}{debug mode, TRUE/FALSE}

\item{configs}{\code{Configs} object for \code{fbseq}}
}
\value{
path to simulated objects
}
\description{
run benchmark methods on simulation objects
}

