% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Function \code{fit}}
\usage{
fit(path, benchmarks = "edgeR", fbseq_methods = c("fullybayes", "ibayes",
  "ebayesFromFullybayes", "ebayesFromStarts", "ebayesFromTruth"),
  priors = c("normal", special_beta_priors()), ncores = detectCores(),
  configs = Configs(), zeronormfactors = F)
}
\arguments{
\item{path}{to directory to save simulations and results}

\item{benchmarks}{benchmark methods to run}

\item{fbseq_methods}{"ebayesFromFullybayes", "ebayesFromStarts", "ebayesFromTruth", "fullybayes", 
or a combination/vector of these. "fullybayes" must come before "ebayesFromFullybayes".}

\item{priors}{hierarchical distributions on the betas. Either a vector, each of whose elements 
is in the slot Configs()@priors, or a list of vectors, where each vector is in the slot Configs@priors.}

\item{ncores}{number of cores for CPU-parallel methods}

\item{configs}{\code{Configs} object for \code{fbseq}}

\item{zeronormfactors}{TRUE/FALSE. If TRUE, starts@h is set to 0.}
}
\value{
path to simulated objects
}
\description{
run benchmark methods on simulation objects
}

