% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_edgeR.R
\name{fit_edgeR}
\alias{fit_edgeR}
\title{Function \code{fit_edgeR}}
\usage{
fit_edgeR(counts, design, group = NULL, ncores = 1)
}
\arguments{
\item{counts}{RNA-seq count datset}

\item{design}{design matrix}

\item{group}{group vector}

\item{ncores}{number of cores for parallel execution}
}
\value{
a list of components of a fit object from \code{edgeR}
}
\description{
fits edgeR
}

