% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit_fbseq.R
\name{fit_fbseq}
\alias{fit_fbseq}
\title{Function \code{fit_fbseq}}
\usage{
fit_fbseq(sim, depth = "fullybayes", prior = "normal", debug = F)
}
\arguments{
\item{sim}{a list, the current simulation object}

\item{depth}{one of "ebayes", "ebayesFromTruth", or "fullybayes"}

\item{prior}{prior distribution on betas}

\item{debug}{debug mode, TRUE/FALSE}
}
\value{
information for a fitted model
}
\description{
fits with fbseq
}

