% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit_Niemi.R
\name{fit_Niemi}
\alias{fit_Niemi}
\title{Function \code{fit_Niemi}}
\usage{
fit_Niemi(counts, design, group, ncores = 1)
}
\arguments{
\item{counts}{RNA-seq count datset}

\item{design}{design matrix}

\item{group}{group vector}

\item{ncores}{number of cores for parallel execution}
}
\value{
\code{rstan} output list
}
\description{
Fit a dataset using the method in Niemi et al.
}

