% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{my.proc.time}
\alias{my.proc.time}
\title{Function \code{my.proc.time}}
\usage{
my.proc.time()
}
\value{
a named numeric vector version of \code{proc.time()}
}
\description{
\code{proc.time()} as a named numeric vector
}

