% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze-plot_calibration.R
\name{plot_calibration}
\alias{plot_calibration}
\title{Function \code{plot_calibration}}
\usage{
plot_calibration(from, to, analysis = analyses(), reps = 1:10)
}
\arguments{
\item{from}{directory with calibration information}

\item{to}{directory to save plots}

\item{analysis}{analysis methods to plot}

\item{reps}{reps to plot}
}
\description{
plot calibration curves using rds files extracted from simulation lists
}

