% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyze-plot_roc.R
\name{plot_roc}
\alias{plot_roc}
\title{Function \code{plot_roc}}
\usage{
plot_roc(from, to)
}
\arguments{
\item{from}{directory of extracted results files}

\item{to}{directory to save plots}
}
\description{
plot roc curves using rds files extracted from simulation lists
}

