% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study-priors_analyze.R
\name{priors_analyze}
\alias{priors_analyze}
\title{Function \code{priors_analyze}}
\usage{
priors_analyze(from, to)
}
\arguments{
\item{from}{to directory to save simulations and results}

\item{to}{output directory}
}
\description{
Plot and summarize results from simulation studies
}

