% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study-priors_init.R
\name{priors_init}
\alias{priors_init}
\title{Function \code{priors_init}}
\usage{
priors_init(path = newdir())
}
\arguments{
\item{path}{to directory to save simulations and results}
}
\value{
path to simulated objects
}
\description{
Initialize the small simulation study in the priors paper
}

