% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyze-roc.R
\name{roc}
\alias{roc}
\title{Function \code{roc}}
\usage{
roc(probs, truth)
}
\arguments{
\item{probs}{probabilities}

\item{truth}{logical or 0/1 vector of classifications}
}
\value{
data frame with an roc curve
}
\description{
function for calculating ROC curves
}

