% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyze-runtime.R
\name{runtime}
\alias{runtime}
\title{Function \code{runtime}}
\usage{
runtime(from, to)
}
\arguments{
\item{from}{directory with simulation lists}

\item{to}{output directory}
}
\value{
runtime table
}
\description{
Runtimes in computation study
}

