% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyze-runtime.R
\name{runtime}
\alias{runtime}
\title{Function \code{runtime}}
\usage{
runtime(path)
}
\arguments{
\item{path}{directory with simulation lists}
}
\value{
runtime table
}
\description{
Runtimes in computation study
}

