% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation_model.R
\name{simulation_model}
\alias{simulation_model}
\title{Function \code{simulation_model}}
\usage{
simulation_model(genes = 35000, libraries = 16)
}
\arguments{
\item{genes}{number of genes}

\item{libraries}{number of libraries}
}
\value{
a list if pertinent scenario information
}
\description{
simulates a scenario using model
}

